deep-research-agent/requirements.txt

langchain
langgraph
langchain-community
tavily-search
openai
python-dotenv
